+++ llm_bioinformatics_research/src/Components/HomeScreen/HomeScreen.css
@@ -1,24 +1,20 @@
-/* General Home Screen */
 .home-screen {
   height: 97vh;
   display: flex;
   flex-direction: column;
 }
 
-/* Main Content */
 .main-content {
   flex-grow: 1;
   padding: 16px;
 }
 
-/* AI Chat Section */
 .ai-chat {
   display: flex;
   flex-direction: column;
   width: 100%;
 }
 
-/* Custom Paper Styling */
 .custom-paper {
   padding: 16px;
   background-color: var(--panel-background); 
@@ -84,7 +80,6 @@
   border: 1px solid var(--panel-border);
 }
 
-/* Dynamic Theme Variables */
 :root {
   --panel-background: #f5f5f558; 
   --panel-border: #b0bec5;
@@ -93,7 +88,6 @@
   --text-color: #33333;
 }
 
-/* Dark Mode Overrides */
 [data-theme='dark'] {
   --panel-background: #2c2c2c; 
   --panel-border: #c3c2c2;
