+++ llm_bioinformatics_research/src/Components/LoginScreen/LoginScreen.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { TextField, Button, Container, Typography, Box, IconButton, Link } from '@mui/material';
 import { Visibility, VisibilityOff } from '@mui/icons-material';
 import { useNavigate, Link as RouterLink } from 'react-router-dom';
-import config from './../config.json';
+import config from '../../config.json';
 import './LoginScreen.css';
 
 function LoginScreen() {
