+++ llm_bioinformatics_research/src/Components/SignupScreen/SignupScreen.js
@@ -4,7 +4,7 @@ import { Visibility, VisibilityOff } from '@mui/icons-material';
 import { useNavigate, Link as RouterLink } from 'react-router-dom';
 import GoogleIcon from '@mui/icons-material/Google';
 import GitHubIcon from '@mui/icons-material/GitHub';
-import config from './../config.json';
+import config from '../../config.json';
 import './SignupScreen.css';
 
 function SignUpScreen() {
