+++ llm_bioinformatics_research/src/Components/UserProfile/UserProfile.css
@@ -1,4 +1,3 @@
-/* App Bar */
  .appBar {
     justify-content: center;
     background-color: #1976d2;
@@ -9,15 +8,12 @@
     text-align: center;
   }
   
-  /* Button Section */
   .button-section {
     margin-top: 16px;
     display: flex;
     justify-content: space-around;
   }
   
-  
-  /* Profile Section */
   .profile-section {
     margin-top: 24px;
     display: flex;
@@ -31,12 +27,10 @@
     margin-bottom: 20px;
   }
    
-  /* Input Fields */
   .input-field {
     width: 80%;
   }
   
-  /* Settings Section */
   .settings-section {
     margin-top: 24px;
     position: relative;
