+++ llm_bioinformatics_research/src/Testing/HomeScreen.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import HomeScreen from '../HomeScreen/HomeScreen';
+import HomeScreen from '../Components/HomeScreen/HomeScreen';
 import '@testing-library/jest-dom/extend-expect';
 import App from '../App';
