+++ llm_bioinformatics_research/src/Testing/LoginScreen.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { MemoryRouter } from 'react-router-dom';
-import LoginScreen from '../LoginScreen/LoginScreen';
+import LoginScreen from '../Components/LoginScreen/LoginScreen';
 import { enableFetchMocks } from 'jest-fetch-mock';
 
 enableFetchMocks();
